re-organize orders for multi reinforcement agent
